+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support for a terminal-style cursor position indicator with a configurable appearance. Requires
   the `alloc` feature to remain enabled. The cursor wrapper is added by calling the `with_cursor`
   method, which can be done before or after any other wrappers are added.
+- Fourth example that uses `embedded-graphics-simulator`: cursor settings, rendering dinamically.
 
 ## [0.1.0-beta.2] - 2026-01-31